+++ src/roq/bitmex/order_entry.cpp
@@ -356,7 +356,7 @@ void OrderEntry::create_order_ack(
       case core::http::Category::SUCCESS: {  // 2xx
         auto body = response.body();
         auto order_item = core::json::Parser::create<json::OrderItem>(body);
-        OrderUpdate{shared_, stream_id_, security_.get_account()}(order_item, trace_info);
+        OrderUpdate{shared_, stream_id_, security_.get_account()}(order_item, trace_info, false);
         break;
       }
       case core::http::Category::CLIENT_ERROR: {
@@ -417,7 +417,7 @@ void OrderEntry::modify_order_ack(
     switch (category) {
       case core::http::Category::SUCCESS: {  // 2xx
         auto order_item = core::json::Parser::create<json::OrderItem>(response.body());
-        OrderUpdate{shared_, stream_id_, security_.get_account()}(order_item, trace_info);
+        OrderUpdate{shared_, stream_id_, security_.get_account()}(order_item, trace_info, false);
         break;
       }
       case core::http::Category::CLIENT_ERROR: {  // 4xx
@@ -479,7 +479,7 @@ void OrderEntry::cancel_order_ack(
       case core::http::Category::SUCCESS: {  // 2xx
         core::json::Buffer buffer(decode_buffer_);
         auto order = core::json::Parser::create<json::Order>(response.body(), buffer);
-        OrderUpdate{shared_, stream_id_, security_.get_account()}(order, trace_info);
+        OrderUpdate{shared_, stream_id_, security_.get_account()}(order, trace_info, false);
         break;
       }
       case core::http::Category::CLIENT_ERROR: {  // 4xx
@@ -537,7 +537,7 @@ void OrderEntry::cancel_all_orders_ack(const core::web::Response &response) {
       case core::http::Category::SUCCESS: {  // 2xx
         core::json::Buffer buffer(decode_buffer_);
         auto order = core::json::Parser::create<json::Order>(response.body(), buffer);
-        OrderUpdate{shared_, stream_id_, security_.get_account()}(order, trace_info);
+        OrderUpdate{shared_, stream_id_, security_.get_account()}(order, trace_info, false);
         break;
       }
       case core::http::Category::CLIENT_ERROR: {  // 4xx
@@ -561,12 +561,12 @@ void OrderEntry::cancel_all_orders_ack(const core::web::Response &response) {
 
 void OrderEntry::operator()(const json::OrderItem &order_item) {
   server::TraceInfo trace_info;
-  OrderUpdate{shared_, stream_id_, security_.get_account()}(order_item, trace_info);
+  OrderUpdate{shared_, stream_id_, security_.get_account()}(order_item, trace_info, false);
 }
 
 void OrderEntry::operator()(const json::Order &order) {
   server::TraceInfo trace_info;
-  OrderUpdate{shared_, stream_id_, security_.get_account()}(order, trace_info);
+  OrderUpdate{shared_, stream_id_, security_.get_account()}(order, trace_info, false);
 }
 
 }  // namespace bitmex